move tests from test/python/quantum_info to test/python/synthesis
