+++ test/python/synthesis/test_local_invariance.py
