+++ test/python/synthesis/xx_decompose/test_circuits.py
