+++ test/python/synthesis/xx_decompose/test_polytopes.py
